# apps/client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KduFoot API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/clubs/search": {
      "get": {
        "tags": [
          "Clubs"
        ],
        "summary": "Search for clubs by name",
        "description": "Searches the local database for football clubs matching the provided name query. This is typically used in autocomplete fields to help users find and link their club. Requires a query string of at least 3 characters to ensure relevance and performance.\n",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search query for the club name (min 3 chars).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of clubs matching the search criteria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "clubs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Club"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - The query is missing or too short."
          }
        }
      }
    },
    "/api/clubs/by-city": {
      "get": {
        "tags": [
          "Clubs"
        ],
        "summary": "Retrieve clubs by city",
        "description": "Returns a list of clubs located in a specific city. Useful for regional filtering and helping users find clubs near their location.\n",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": true,
            "description": "Name of the city to filter by.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of clubs found in the specified city.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "clubs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Club"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - The city parameter is missing."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Club": {
        "type": "object",
        "description": "Represents a football club or organization identified by its SIRET number.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Internal unique identifier for the club."
          },
          "siret": {
            "type": "string",
            "description": "Unique 14-digit SIRET identifier."
          },
          "name": {
            "type": "string",
            "description": "Full official name of the club."
          },
          "city": {
            "type": "string",
            "description": "City where the club is located."
          },
          "address": {
            "type": "string",
            "description": "Full street address of the club's headquarters or headquarters."
          },
          "zip": {
            "type": "string",
            "description": "Zip or postal code."
          },
          "latitude": {
            "type": "number",
            "description": "Geographical latitude for mapping."
          },
          "longitude": {
            "type": "number",
            "description": "Geographical longitude for mapping."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ]
}